+++ server/src/main/java/org/realityforge/replicant/server/transport/ChannelMetaData.java
@@ -1,8 +1,13 @@
 package org.realityforge.replicant.server.transport;
 
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.Objects;
+import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.UnmodifiableView;
 
 @SuppressWarnings( "WeakerAccess" )
 public final class ChannelMetaData
@@ -62,6 +67,8 @@ public final class ChannelMetaData
   private final boolean _external;
   @Nonnull
   private final ChannelMetaData[] _requiredTypeChannels;
+  @Nonnull
+  private final Set<ChannelMetaData> _dependentChannels = new HashSet<>();
 
   public ChannelMetaData( final int channelId,
                           @Nonnull final String name,
@@ -97,6 +104,7 @@ public final class ChannelMetaData
         throw new IllegalArgumentException( "Specified RequiredTypeChannel " + requiredTypeChannel.getName() +
                                             " is not a type channel" );
       }
+      requiredTypeChannel._dependentChannels.add( this );
     }
   }
 
@@ -171,5 +179,13 @@ public final class ChannelMetaData
   public ChannelMetaData[] getRequiredTypeChannels()
   {
     return _requiredTypeChannels;
+  }
+
+  @Contract( pure = true )
+  @Nonnull
+  @UnmodifiableView
+  public Set<ChannelMetaData> getDependentChannels()
+  {
+    return Collections.unmodifiableSet( _dependentChannels );
   }
 }