+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Remove `CacheType.EXTERNAL` as no longer in use.
+* Cascade cache invalidations to dependent cached type graphs to avoid failure scenario when a chain of cached graphs as an invalidation in internal node of chain.
 
 ### [v6.126](https://github.com/replicant4j/replicant/tree/v6.126) (2022-09-27) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.125...v6.126)